requirements.txt
# Core
torch>=2.2
torchvision
transformers>=4.42
accelerate>=0.30
datasets>=2.19
peft>=0.11
trl>=0.9
pillow>=10.2
pyyaml>=6.0

# Evaluation & metrics
evaluate>=0.4
bert-score>=0.3
rouge-score>=0.1

# UI
gradio>=4.0

# Optional (only if you actually use GCS anywhere)
google-cloud-storage>=2.16
